+++ setup_complete.sh
@@ -1,1 +1,0 @@
-# Paste the ENTIRE content from 1786463122357.txt here